chore: Tests variables should be global
Part of #13

If they are declared inside the describe they are bind to its context
and thus lost. The test case and before* are run outside of the describe
callback and don't have access to it. That is why variables shared
between multiple test cases must be declared in global context.