chore: Only expose necessary .h
Part of #43

include dir should only expose test.h, utilities.h and assert.h

Others header are hidden inside internals directory and should not be
used externaly (hence the name + namespace).

Only those 3 files are exposed as they directly contains all the tools
for tests.

In the case of utilities it is just some functions used internally but
with a wide range of use and not just the tests. It seems ok to expose
them